Update README with more information.
This is actually a bit more and a bit less, since it no longe rlinks
to #3#. I've made the decision to keep a small part of my repository
closed source, so linking to it as a runnable example is probably not
the best idea!

If anyone else has fully open source examples, I'll be glad to add
them.