Fix spawnCreep JavaScript.
This partially reverts
https://github.com/daboross/screeps-in-rust-via-wasm/commit/e7be1e65955f4d50b656c2d60848312c572ed664.

I completely missed the change, but it's an important one.
`js_unwrap!()` cannot be used with javascript bodies which contain
multiple statements and "return", as it unconditionally inserts a
return statement at the start of the body. "return var body = ..." is a
syntax error, but since we don't have any test interacting with actual
JavaScript, it was not caught by the CI.