Fix spawn_with_options execution.
This seems to have been broken from the very first time it was
created... I'm fixing it now since there was something else broken
here, and it might as well be fixed. `spawnCreep(@{body}...` wouldn't
work since at this point body is an array of integer codes, not part
names.