Fix `LocalRoomPosition`'s serialize implementation
This is a slightly complicated change, mostly because I want to retain
compatibility with the old format, and to retain its behavior in binary
formats. There are now two possible serialized forms: one which stores
room_name as a string, and one which stores it as two integers.

When serializing, it will choose the one with the room name for
readable formats like sending to JavaScript, or JSON, or messagepack,
and choose the one with integers when using pure-binary ones like
bincode.

When deserializing, it'll deserialize the one with integers when using
a pure-binary format, and try both for human readable formats (to keep
compatibility with old behavior, where the one with integers was
always serialized)