Remove remaining usages of get_from_js!
As far as macros go, I don't think this one improved the amount of code
we have significantly, and it had a kind of opaque syntax. Invocations
already had to supply the name and type, all this did was replace
`js_unwrap!` and `pub fn` with `get_from_js!` and a different syntax.