fix(local_cost_matrix): added asserts that x and y are valid coordinates
Before, we just relied on the bounds checking the index that pos_as_idx spits out, but that allows
invalid coordinates like (0, 201) to be passed in. This fixes that. As a result of the bounds
checking from the asserts, we can skip the checking that goes on in []::index, and just use the
unchecked versions instead.