refactor(local_cost_matrix): get and set now use indexing, indexing uses unchecked internally
With the (u8, u8) indexing impl, we can rewrite get and set in terms of them. And in the indexing
impl, because we check that x and y are in bounds, we're guaranteed that the index from pos_as_idx
will be in bounds as well, so we can just use the unchecked accessors instead.