LocalCostMatrix improvements (#347)
* refactor(local_cost_matrix): changed LocalCostMatrix to use a fixed-size array internally

Instead of using Vec<u8>, LocalCostMatrix now holds a [u8; 2500] directly. Also replaced the
Into<Vec<u8>> for LocalCostMatrix with From<LocalCostMatrix> for Vec<u8>, to be idiomatic.

* feat(local_cost_matrix): added indexing implementations for (u8, u8)

LocalCostMatrix's can now be indexed by (u8, u8) pairs.

* feat(local_cost_matrix): added unchecked get and set

* fix(local_cost_matrix): added asserts that x and y are valid coordinates

Before, we just relied on the bounds checking the index that pos_as_idx spits out, but that allows
invalid coordinates like (0, 201) to be passed in. This fixes that. As a result of the bounds
checking from the asserts, we can skip the checking that goes on in []::index, and just use the
unchecked versions instead.

* feat(local_cost_matrix): added indexing impls for Position

* refactor(local_cost_matrix): get and set now use indexing, indexing uses unchecked internally

With the (u8, u8) indexing impl, we can rewrite get and set in terms of them. And in the indexing
impl, because we check that x and y are in bounds, we're guaranteed that the index from pos_as_idx
will be in bounds as well, so we can just use the unchecked accessors instead.