SparseCostMatrix (#348)
* refactor(local_cost_matrix): changed LocalCostMatrix to use a fixed-size array internally

Instead of using Vec<u8>, LocalCostMatrix now holds a [u8; 2500] directly. Also replaced the
Into<Vec<u8>> for LocalCostMatrix with From<LocalCostMatrix> for Vec<u8>, to be idiomatic.

* feat(local_cost_matrix): added indexing implementations for (u8, u8)

LocalCostMatrix's can now be indexed by (u8, u8) pairs.

* feat(local_cost_matrix): added unchecked get and set

* fix(local_cost_matrix): added asserts that x and y are valid coordinates

Before, we just relied on the bounds checking the index that pos_as_idx spits out, but that allows
invalid coordinates like (0, 201) to be passed in. This fixes that. As a result of the bounds
checking from the asserts, we can skip the checking that goes on in []::index, and just use the
unchecked versions instead.

* feat(local_cost_matrix): added indexing impls for Position

* refactor(local_cost_matrix): get and set now use indexing, indexing uses unchecked internally

With the (u8, u8) indexing impl, we can rewrite get and set in terms of them. And in the indexing
impl, because we check that x and y are in bounds, we're guaranteed that the index from pos_as_idx
will be in bounds as well, so we can just use the unchecked accessors instead.

* feat(cost_matrix): added a SparseCostMatrix struct, added LocalCostMatrix::{iter, iter_mut}

SparseCostMatrix is the sparse counterpart to the LocalCostMatrix. Also added methods for generating
iterators over both the sparse and dense versions.

* feat(sparse_cost_matrix): added From<CostMatrix> implementation

* perf(sparse_cost_matrix): refactored assert in From<CostMatrix> to only run once

* refactor(local/cost_matrix.rs): refactored index -> (x,y) into a separate function

* feat(local/cost_matrix.rs): added methods for merging cost matrices

Can now merge the two cost matrix types into each other; forms the basis for possible layered cost
matrix behaviour, builder patterns, etc.

* refactor(cost_matrix_iter): rewrote the iter methods to give &u8 values back instead of u8

Some users might only care about a few values, why deref before it's needed? They can always chain
.copied() after if needed.

BREAKING CHANGE: .iter() returns impl Iterator<Item = ((u8, u8), &u8)> instead of impl Iterator<Item
= ((u8, u8), u8)> for Local and SparseCostMatrix.

* fix(local/cost_matrix.rs): fixed impls that broke due to iter output change

* refactor(local_cost_matrix): removed casts of Position::x and Position::y to u8

Now that the position change is merged, these casts are unnecessary.

* style(local/cost_matrix.rs): removed old commented-out code

* docs(sparse_cost_matrix): added comment explaining why 0 entries are filtered out

* style(local/cost_matrix.rs): removed more std-web code