refactor(pathfinder.rs): refactored non-callback SearchOptions fields into an inner structure
Instead of having all the non-generic SearchOptions fields live in the structure, have an inner
structure with a derived Serialize implementation for ease of conversion. Also reduces code needed
for changing the callback, or creating a new SearchOptions.

BREAKING CHANGE: SearchOptions::new() now requires a callback be given in the argument.