Wrap raw store functions
The get_ functions on Store can return null if the resource type is
not valid for that store. This can cause unexpected panics if, e.g.
a creep tries to check its store when all of its Carry parts have
been destroyed.

Renamed the raw functions with _internal, following the pattern used
elsewhere, made them accept Option<i/u32>s, and had the public
functions unwrap those options. 0 is used as a default value, which
makes conceptual sense for all funcitons: a store that can't take a
resource has no capacity for it and currently stores none.

Resolves #429